env and factory refactoring
