+++ TrafficCapture/dockerSolution/src/main/docker/docker-compose.yml
@@ -97,7 +97,7 @@ services:
       - sharedReplayerOutput:/shared-replayer-output
       - ./migrationConsole/lib/console_link/services.yaml:/etc/migration_services.yaml
       # this is a convenience thing for testing -- it should be removed before this makes it to prod.
-      # - ./migrationConsole/lib/console_link:/root/lib/console_link
+      - ./migrationConsole/lib/console_link:/root/lib/console_link
       - ~/.aws:/root/.aws
     environment:
       - MIGRATION_KAFKA_BROKER_ENDPOINTS=kafka:9092
