+++ TrafficCapture/dockerSolution/src/main/docker/migrationConsole/lib/console_link/console_link/environment.py
@@ -2,8 +2,8 @@ import logging
 from typing import Optional
 from console_link.models.cluster import Cluster
 from console_link.models.metrics_source import MetricsSource
-from console_link.middleware.metrics import get_metrics_source
-from console_link.middleware.backfill import get_backfill
+from console_link.middleware.factories import get_metrics_source
+from console_link.middleware.factories import get_backfill
 from console_link.models.backfill_base import Backfill
 import yaml
 from cerberus import Validator
