+++ TrafficCapture/dockerSolution/src/main/docker/migrationConsole/lib/console_link/console_link/middleware/exit_code.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class ExitCode(Enum):
+    SUCCESS = 0
+    FAILURE = 1
